Live Chat: Prevent scroll bleed.
In this branch of a branch, I've used JS to prevent the main body content from scrolling when your mouse is over the main chat panel.

It's not super pretty, but it feels solid so far. I think we should merge it into `try/live-chat`, and test rigorously, but would still appreciate eyes on it first.